pyproject.toml
[project]
name = "infrastructure-deployments"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "paho-mqtt>=2.1.0",
    "pandas>=2.3.3",
    "python-dotenv>=1.2.1",
    "ual",
]

[tool.uv.sources]
ual = { git = "https://github.com/urban-air-lab/common" }
